requirements.txt
Flask==1.0.*

# Application server for both development and production.
gunicorn==19.8.*

# libs
iso-639==0.4.5
nltk==3.3
lxml==4.2.3

# Testing and static analysis.
pytest==3.5.*
pytest-cov==2.5.*
flake8==3.5.*

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.9.7

# Extensions.
Flask-Login==0.4.*